requirements.txt
requests~=2.31.0
pytest~=7.4.3
PyYAML~=6.0.1
selenium~=4.15.2
PyMySQL~=1.1.0